NC updates, revert Moldavian change
